reactive_graph_behaviour_service_api::entity_behaviour_manager

Trait EntityBehaviourManager

pub trait EntityBehaviourManager:
    Send
    + Sync
    + Lifecycle {
    // Required methods
    fn add_behaviours(&self, entity_instance: ReactiveEntity);
    fn add_behaviour(
        &self,
        entity_instance: ReactiveEntity,
        behaviour_ty: &BehaviourTypeId,
    );
    fn remove_behaviour(
        &self,
        entity_instance: ReactiveEntity,
        behaviour_ty: &BehaviourTypeId,
    );
    fn remove_behaviours(&self, entity_instance: ReactiveEntity);
    fn remove_behaviours_by_id(&self, id: &Uuid);
    fn remove_behaviours_by_behaviour(&self, behaviour_ty: &BehaviourTypeId);
    fn has(
        &self,
        entity_instance: ReactiveEntity,
        behaviour_ty: &BehaviourTypeId,
    ) -> bool;
    fn get_all(&self, entity_instance: ReactiveEntity) -> Vec<BehaviourTypeId>;
    fn get_instances_by_behaviour(
        &self,
        ty: &BehaviourTypeId,
    ) -> Vec<ReactiveEntity>;
    fn connect(
        &self,
        entity_instance: ReactiveEntity,
        behaviour_ty: &BehaviourTypeId,
    ) -> Result<(), BehaviourTransitionError>;
    fn disconnect(
        &self,
        entity_instance: ReactiveEntity,
        behaviour_ty: &BehaviourTypeId,
    ) -> Result<(), BehaviourTransitionError>;
    fn reconnect(
        &self,
        entity_instance: ReactiveEntity,
        behaviour_ty: &BehaviourTypeId,
    ) -> Result<(), BehaviourTransitionError>;
}

Required Methods

fn add_behaviours(&self, entity_instance: ReactiveEntity)

Adds all behaviours to the given reactive entity instance.

fn add_behaviour( &self, entity_instance: ReactiveEntity, behaviour_ty: &BehaviourTypeId, )

Creates and adds the given behaviour to the given reactive entity instance.

fn remove_behaviour( &self, entity_instance: ReactiveEntity, behaviour_ty: &BehaviourTypeId, )

Removes the given behaviour from the given reactive entity instance.

fn remove_behaviours(&self, entity_instance: ReactiveEntity)

Removes all behaviours from the given reactive entity instance.

fn remove_behaviours_by_id(&self, id: &Uuid)

Removes all behaviours from the reactive entity instance with the given id.

fn remove_behaviours_by_behaviour(&self, behaviour_ty: &BehaviourTypeId)

Removes all behaviours of the given behaviour type.

fn has( &self, entity_instance: ReactiveEntity, behaviour_ty: &BehaviourTypeId, ) -> bool

Returns true, if the entity instance has the given behaviour.

fn get_all(&self, entity_instance: ReactiveEntity) -> Vec<BehaviourTypeId>

Returns the behaviours of the given entity instance.

fn get_instances_by_behaviour( &self, ty: &BehaviourTypeId, ) -> Vec<ReactiveEntity>

Returns the entity instances with the given behaviour.

fn connect( &self, entity_instance: ReactiveEntity, behaviour_ty: &BehaviourTypeId, ) -> Result<(), BehaviourTransitionError>

Connect

fn disconnect( &self, entity_instance: ReactiveEntity, behaviour_ty: &BehaviourTypeId, ) -> Result<(), BehaviourTransitionError>

Disconnect

fn reconnect( &self, entity_instance: ReactiveEntity, behaviour_ty: &BehaviourTypeId, ) -> Result<(), BehaviourTransitionError>

Reconnect